% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_data.R
\name{draft_data}
\alias{draft_data}
\title{Draft data function
This function gets data from all players drafted in the years specified, and returns a dataframe}
\usage{
draft_data(startYear, endYear)
}
\arguments{
\item{startYear}{= First year you want draft data from, endYear = last year you want draft data from. BOth values must be entered.}
}
\description{
Draft data function
This function gets data from all players drafted in the years specified, and returns a dataframe
}
\examples{
#Returns regular season statistics from all players drafted between 2012 and 2016
draft_data(2012, 2016)

}
\keyword{icehockey,}
\keyword{stats}
