% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_extract.R
\name{draft_extract}
\alias{draft_extract}
\title{Draft extract function
Nested function that scrapes and extracts the data from the Eliteprospects website. Called automatically by draft_data function in hockeyStats package}
\usage{
draft_extract(year)
}
\arguments{
\item{year}{= sequence of years for whcih draft data is to be extracted and returned}
}
\description{
Draft extract function
Nested function that scrapes and extracts the data from the Eliteprospects website. Called automatically by draft_data function in hockeyStats package
}
\examples{
draft_extract(seq(2012, 2016, 1))

}
\keyword{icehockey,}
\keyword{stats}
