% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_FUN.R
\name{round_FUN}
\alias{round_FUN}
\title{Function to get the individual draft rounds frp, each draft year. Called by the draft_extract function in the hockeyStats package
This function returns list of the numbers of the draft rounds from each draft year}
\usage{
round_FUN(no_rounds, draft_rounds)
}
\arguments{
\item{no_rounds}{= number of draft rounds that particular year. draft_rounds = length of individual draft rounds}
}
\description{
Function to get the individual draft rounds frp, each draft year. Called by the draft_extract function in the hockeyStats package
This function returns list of the numbers of the draft rounds from each draft year
}
\examples{
roundFun(no_rounds, draft_rounds)

}
\keyword{icehockey,}
\keyword{stats}
